[JAX] Increase numerical tolerances of tests in preparation for an XLA:CPU vectorization change.
An upcoming change to XLA:CPU alters the LLVM compilation pipeline to avoid relying as heavily on the superword-level parallelism (SLP) vectorizer, which has been a frequent source of bugs.

However, changes to LLVM vectorization can lead to numerical differences in tests. This change increments the numerical tolerances of a number of tests, in preparation for landing the XLA:CPU change.

PiperOrigin-RevId: 401256780