Fix weight calculations in mixup.
Originally, this code drew two weights from the dirichlet distribution, which only has support over the simplex, so the two weights added to 1.

However, when moving to tf.data, we have neither a stateless beta or a stateless delta, so we used a stateless gamma to get our results. However, gamma _does_ have support beyond the simplex, so our weights were summing to greater than 1.

PiperOrigin-RevId: 445138329