Remove shuffling in `tfds.load`.
In the multihost setting, each host independently loads 1/num_hosts data and then this data is cached. After pre-emption, a different subset is loaded. Without pre-emption, we may be not loading large portions of the data since each host shuffles differently.

The alternative is to pass a seed to `tfds.load` to make file shuffling deterministic across hosts and pre-emptions:

https://www.tensorflow.org/datasets/determinism#get_deterministic_file_shuffling

PiperOrigin-RevId: 446716275