[NumPy] Disable init2winit test that fails under NumPy 1.23.
I am unsure what the right fix is and the authors of the code are probably the best people to figure that out, but note that under NumPy 1.21 the test issues the following warnings which are probably of immediate relevance given NumPy 1.23 drops the backward-compatibility behavior in question:

init2winit/model_lib/metrics.py:176: FutureWarning: Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will result either in an error or a different result.
  valid_targets = targets[weights > 0]
init2winit/model_lib/metrics.py:201: FutureWarning: Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will result either in an error or a different result.
  valid_targets = targets[weights > 0]

PiperOrigin-RevId: 466206773