Fix timing code in train loop.
utils.timed does not account for jax's async dispatch. It works correctly for timing eval code because it waits on the entire eval loop / result. Originally, we wanted to simplify the timing code and isolate timing for just update_pmapped.

However, because of async dispatch, in order to make it work correctly in the train loop, we would need to add a sync barrier to wait on the result of update_pmapped.

Given that we have to choose between correctness and adding a sync barrier, we propose to revert to the old style of timing the train loop where we time between eval calls.

Pros:
- Not wrong
- Correctly accounts for train + data + other

Cons:
- Becomes eval timing and "everything else" timing
- Slightly more complex timing logic

NOTE: we may be able to time asynchronously via pure jax callbacks.
PiperOrigin-RevId: 477440658