Refactoring the Criteo model and dataset. The model used to have multiple embeddings (one per feature) but now we just use a single shared embedding, and we concatenate the feature column index to the feature string before hashing to generate embedding indices.
This updates the Criteo dataset to work with the raw, externally available data.

This fixes a bug we had in init2winit where we were splitting the first half of day 23 data into validation/test, when that should only be the test set.

Removes `use_half_last_criteo_day_as_test` and assumes it is true.

PiperOrigin-RevId: 489982968