Update PRNG key indexing to support custom PRNG
This change will not modify the code's behavior, but will make the code compatible with JAX's new custom PRNG, which will soon be the default; see https://github.com/google/jax/issues/9263.

Note however that the pattern used here (indexing into key data in order to extract an integer seed) is not recommended because it discards half of the key's bits and may inadvertently lead to non-independent or even identical pseudo-random streams. For example, `seed = jax.random.PRNGKey(N)[0]` will return the same seed for any 32-bit integer `N`.

Instead, we recommend using `seed = jax.random.bits(key)`. Because this option will change the seed output and thus change the behavior of the code, we leave it as a TODO for package maintainers to consider and implement on a case-by-case basis.

PiperOrigin-RevId: 533300858