Save global step after optimizer update.
This way, if evals NaN and raise, you can save the last checkpoint before the nan occurs.

This CL modifies the training loop to add one more `_save` call -- so that the last step's checkpoint is also saved (if requested by the config). This only modifies the checkpoints in the `$train_dir/checkpoints/` directory which correspond to `checkpoint_steps`, not the checkpoint in the `$train_dir` directly (which is saved during `eval_steps`).

PiperOrigin-RevId: 539812769