Add a registry for training algorithms.
This change introduces a `training_algorithms.py` module to register different training algorithm classes by name. The `main.py` script is updated to use a new `--training_algorithm` flag and the `config.training_algorithm` to select which algorithm to use, defaulting to `optax_training_algorithm`. The selected class is then passed to the `Trainer`.

PiperOrigin-RevId: 797408283