[pmap] Fix errors introduced by the new `jax.pmap` implementation.
The new `jax.pmap` is much more strict about data placement. There were (non-deterministic) cases where the rng seed might be placed with a different device ordering than e.g., the params.

This fix explicitly sets the device order for rng so that it should match the device orders for other args to the pmapped step function.

As an aside, we also remove the pmap to create a synchronized seed.

PiperOrigin-RevId: 840372056